Clear all /W4 warnings; drop the About box tagline
All six configurations now build warning-free.

C4100, unreferenced parameters:
- MonitorEnumProc's signature is fixed by EnumDisplayMonitors, so the two
  unused parameters stay and only their names go.
- HandleWindowBoundsChange never used its HWND; it works off MonitorDat and
  pDisplaySpecificData, exactly as HandleTaskBarChange does. The parameter is
  removed rather than silenced, and the three call sites follow.

C4244, narrowing conversions:
- SnowFlake's float coordinates become pixel indices via explicit casts.
  Deliberately behaviour-preserving: the casts truncate toward zero rather
  than flooring, so a flake in the left margin at x in (-1, 0) still lands on
  column 0 instead of failing the bounds check below it. That is a pre-existing
  one-column bias, now commented rather than quietly changed.
- TBM_GETPOS returns LRESULT; the trackbar ranges are 5..50 and -5..5, so the
  narrowing to int cannot lose anything.

Also removes the "Make rain and snow more awesome" line from the About box.
The controls below it move up by the 18 units it occupied and the dialog
loses the same height, so the remaining spacing is unchanged.

+++ Src/let-it-rain/OptionDialog.cpp
@@ -138,7 +138,9 @@ LRESULT CALLBACK OptionsDialog::DialogProc(const HWND hWnd, const UINT message, 
 	case WM_HSCROLL:
 		if (reinterpret_cast<HWND>(lParam) == GetDlgItem(hWnd, IDC_SLIDER))
 		{
-			const int pos = SendMessage(GetDlgItem(hWnd, IDC_SLIDER), TBM_GETPOS, 0, 0);
+			// TBM_GETPOS returns the position in an LRESULT; this trackbar's
+			// range is 5..50, so narrowing to int cannot lose anything.
+			const int pos = static_cast<int>(SendMessage(GetDlgItem(hWnd, IDC_SLIDER), TBM_GETPOS, 0, 0));
 			for (CallBackWindow* subscriber : subscribers)
 			{
 				subscriber->UpdateParticleCount(pos);
@@ -146,7 +148,8 @@ LRESULT CALLBACK OptionsDialog::DialogProc(const HWND hWnd, const UINT message, 
 		}
 		else if (reinterpret_cast<HWND>(lParam) == GetDlgItem(hWnd, IDC_SLIDER2))
 		{
-			const int pos = SendMessage(GetDlgItem(hWnd, IDC_SLIDER2), TBM_GETPOS, 0, 0);
+			// Same again; this one's range is -5..5.
+			const int pos = static_cast<int>(SendMessage(GetDlgItem(hWnd, IDC_SLIDER2), TBM_GETPOS, 0, 0));
 			for (CallBackWindow* subscriber : subscribers)
 			{
 				subscriber->UpdateWindDirection(pos);
